requirements.txt
Django>=1.11.3
Markdown>=2.6.8
django-crispy-forms>=1.6.1
django-cleanup>=1.0.0
django-htmlmin>=0.10.0
django-import-export>=0.5.1
django-cookie-law>=1.0.13
